recognition/requirements.txt
kafka-python==2.0.2
Pillow==8.3.1
mido==1.2.10